pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "psrcdataportal"
dynamic = ["version"]
description = "Tools to export tables and spatial layers to ArcGIS Online Data Portal"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Puget Sound Regional Council", email = "datateam@psrc.org"}
]
maintainers = [
    {name = "Puget Sound Regional Council", email = "datateam@psrc.org"}
]
keywords = [
    "gis",
    "arcgis",
    "data-portal",
    "spatial-data",
    "database",
    "export",
    "psrc",
    "transportation",
    "planning"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Database",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.3.0",
    "geopandas>=0.10.0",
    "sqlalchemy>=1.4.0",
    "pyodbc>=4.0.0",
    "pyyaml>=5.4.0",
    "shapely>=1.8.0",
    "arcgis>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
    "pre-commit>=2.0",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
    "myst-parser>=0.15",
]
test = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "pytest-mock>=3.0",
]

[project.urls]
Homepage = "https://github.com/psrc/psrcdataportal"
Documentation = "https://psrcdataportal.readthedocs.io/"
Repository = "https://github.com/psrc/psrcdataportal.git"
"Bug Tracker" = "https://github.com/psrc/psrcdataportal/issues"

[project.scripts]
psrc-export = "psrcdataportal.cli:main"

[tool.setuptools]
packages = ["psrcdataportal"]
include-package-data = true

[tool.setuptools.package-data]
psrcdataportal = ["config/*.yaml", "config/*.yml"]

[tool.setuptools_scm]
write_to = "psrcdataportal/_version.py"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["psrcdataportal"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "arcgis.*",
    "arcpy.*",
    "geopandas.*",
    "pyodbc.*",
    "to_SpatiallyEnabledDataFrame.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["psrcdataportal"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
